solution Input 통일 + 수식 입력
